changed RunsResultActivity to host fragment
