+++ app/src/main/res/layout/single_frag_container_act.xml
@@ -3,13 +3,3 @@
     android:id="@+id/single_frag_container"
     android:layout_width="match_parent"
     android:layout_height="match_parent" />
-
-    <!--<android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"--><!--xmlns:app="http://schemas.android.com/apk/res-auto"--><!--android:layout_width="match_parent"--><!--android:layout_height="match_parent">-->
-
-    <!--<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"--><!--android:id="@+id/single_frag_container"--><!--android:layout_width="match_parent"--><!--android:layout_height="match_parent">-->
-
-    <!--</FrameLayout>-->
-
-    <!--<android.support.design.widget.FloatingActionButton--><!--android:id="@+id/add_fab"--><!--android:layout_width="wrap_content"--><!--android:layout_height="wrap_content"--><!--android:layout_margin="16dp"--><!--android:src="@drawable/ic_add"--><!--app:layout_anchor="@id/single_frag_container"--><!--app:layout_anchorGravity="bottom|right|end" />-->
-
-    <!--</android.support.design.widget.CoordinatorLayout>-->
